Fixed router issues with back end.
